requirements.txt
Flask==3.1.1
Werkzeug==3.1.3
colorama==0.4.6
python-dotenv==1.0.1
